Add GetSideChainDividends in aedpos contract.
