+++ contract/AElf.Contracts.Consensus.AEDPoS/AEDPoSContract_ProcessConsensusInformation.cs
@@ -163,28 +163,6 @@ namespace AElf.Contracts.Consensus.AEDPoS
             ClearExpiredRandomNumberTokens();
         }
 
-        private void ReleaseSideChainDividendsPool()
-        {
-            if (State.TokenHolderContract.Value == null) return;
-            var scheme = State.TokenHolderContract.GetScheme.Call(Context.Self);
-            var isTimeToRelease =
-                (Context.CurrentBlockTime - State.BlockchainStartTimestamp.Value).Seconds
-                .Div(State.PeriodSeconds.Value) > scheme.Period - 1;
-            Context.LogDebug(() => "ReleaseSideChainDividendsPool Information:\n" +
-                                   $"CurrentBlockTime: {Context.CurrentBlockTime}\n" +
-                                   $"BlockChainStartTime: {State.BlockchainStartTimestamp.Value}\n" +
-                                   $"PeriodSeconds: {State.PeriodSeconds.Value}\n" +
-                                   $"Scheme Period: {scheme.Period}");
-            if (isTimeToRelease)
-            {
-                Context.LogDebug(() => "Ready to release side chain dividends pool.");
-                State.TokenHolderContract.DistributeProfits.Send(new DistributeProfitsInput
-                {
-                    SchemeManager = Context.Self
-                });
-            }
-        }
-
         private void ProcessNextTerm(Round nextRound)
         {
             RecordMinedMinerListOfCurrentRound();
