+++ contract/AElf.Contracts.TokenHolder/TokenHolderContract.cs
@@ -79,7 +79,7 @@ namespace AElf.Contracts.TokenHolder
                 SchemeId = scheme.SchemeId
             }).Details.Single();
             var lockedAmount = detail.Shares;
-            State.ProfitContract.RemoveBeneficiary.Send(new RemoveBeneficiaryInput 
+            State.ProfitContract.RemoveBeneficiary.Send(new RemoveBeneficiaryInput
             {
                 SchemeId = scheme.SchemeId,
                 Beneficiary = input.Beneficiary
@@ -87,7 +87,7 @@ namespace AElf.Contracts.TokenHolder
             if (lockedAmount > input.Amount &&
                 input.Amount != 0) // If input.Amount == 0, means just remove this beneficiary.
             {
-                State.ProfitContract.AddBeneficiary.Send(new AddBeneficiaryInput 
+                State.ProfitContract.AddBeneficiary.Send(new AddBeneficiaryInput
                 {
                     SchemeId = scheme.SchemeId,
                     BeneficiaryShare = new BeneficiaryShare
@@ -148,6 +148,7 @@ namespace AElf.Contracts.TokenHolder
             {
                 distributeProfitsInput.AmountsMap.Add(input.AmountsMap);
             }
+
             State.ProfitContract.DistributeProfits.Send(distributeProfitsInput);
             scheme.Period = scheme.Period.Add(1);
             State.TokenHolderProfitSchemes[input.SchemeManager] = scheme;
@@ -227,7 +228,8 @@ namespace AElf.Contracts.TokenHolder
             }).Amount;
 
             var lockId = State.LockIds[input][Context.Sender];
-            Assert(State.LockTimestamp[lockId].AddMinutes(scheme.MinimumLockMinutes) < Context.CurrentBlockTime, "Cannot withdraw.");
+            Assert(State.LockTimestamp[lockId].AddMinutes(scheme.MinimumLockMinutes) < Context.CurrentBlockTime,
+                "Cannot withdraw.");
 
             State.TokenContract.Unlock.Send(new UnlockInput
             {
@@ -285,7 +287,8 @@ namespace AElf.Contracts.TokenHolder
             return scheme;
         }
 
-        private void UpdateTokenHolderProfitScheme(ref TokenHolderProfitScheme scheme, Address manager, bool updateSchemePeriod)
+        private void UpdateTokenHolderProfitScheme(ref TokenHolderProfitScheme scheme, Address manager,
+            bool updateSchemePeriod)
         {
             if (scheme.SchemeId != null && !updateSchemePeriod) return;
             var originSchemeId = State.ProfitContract.GetManagingSchemeIds.Call(new GetManagingSchemeIdsInput
