+++ protobuf/aedpos_contract.proto
@@ -94,6 +94,11 @@ service AEDPoSContract {
     rpc GetRandomHash (google.protobuf.Int64Value) returns (aelf.Hash) {
         option (aelf.is_view) = true;
     }
+    
+    // Side Chain Dividends Pool.
+    rpc GetSideChainDividends (google.protobuf.Int64Value) returns (SideChainDividends) {
+        option (aelf.is_view) = true;
+    }
 }
 
 message InitialAElfConsensusContractInput {
@@ -250,6 +255,10 @@ message LatestPubkeyToTinyBlocksCount {
 message ContributeToSideChainDividendsPoolInput {
     string symbol = 1;
     int64 amount = 2;
+}
+
+message SideChainDividends {
+    map<string, int64> value = 1;
 }
 
 // Events
